refactor: simplify build and packaging of carousel component
