+++ packages/components/carousel/bs-carousel-control-prev.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsCarouselControlPrevCss } from './bs-carousel-control-prev.css.js';
+import { BsCarouselControlPrevCss } from './css/bs-carousel-control-prev.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsCarouselControlPrev extends LitElement {
