+++ packages/components/carousel/css/bs-carousel-control-next.css.js
