+++ packages/components/carousel/css/bs-carousel-control-prev.css.js
