+++ packages/components/carousel/css/bs-carousel-indicator.css.js
