+++ packages/components/carousel/css/bs-carousel-indicators.css.js
