+++ packages/components/carousel/css/bs-carousel-item.css.js
