+++ packages/components/carousel/css/bs-carousel.css.js
