+++ packages/components/carousel/css/index.js
@@ -1,15 +1,8 @@
 
-export { BsCarousel } from './bs-carousel.js';
 export { BsCarouselCss } from './bs-carousel.css.js';
-export { BsCarouselItem } from './bs-carousel-item.js';
 export { BsCarouselItemCss } from './bs-carousel-item.css.js';
-export { BsCarouselCaption } from './bs-carousel-caption.js';
 export { BsCarouselCaptionCss } from './bs-carousel-caption.css.js';
-export { BsCarouselIndicator } from './bs-carousel-indicator.js';
 export { BsCarouselIndicatorCss } from './bs-carousel-indicator.css.js';
-export { BsCarouselIndicators } from './bs-carousel-indicators.js';
 export { BsCarouselIndicatorsCss } from './bs-carousel-indicators.css.js';
-export { BsCarouselControlPrev } from './bs-carousel-control-prev.js';
 export { BsCarouselControlPrevCss } from './bs-carousel-control-prev.css.js';
-export { BsCarouselControlNext } from './bs-carousel-control-next.js';
 export { BsCarouselControlNextCss } from './bs-carousel-control-next.css.js';
