+++ packages/components/carousel/index.js
@@ -1,0 +1,8 @@
+export * from './css'
+export { BsCarousel } from './bs-carousel.js';
+export { BsCarouselItem } from './bs-carousel-item.js';
+export { BsCarouselCaption } from './bs-carousel-caption.js';
+export { BsCarouselIndicator } from './bs-carousel-indicator.js';
+export { BsCarouselIndicators } from './bs-carousel-indicators.js';
+export { BsCarouselControlPrev } from './bs-carousel-control-prev.js';
+export { BsCarouselControlNext } from './bs-carousel-control-next.js';
